+++ src/main/java/me/petrolingus/electricfieldlines/Application.java
@@ -7,11 +7,13 @@ import javafx.stage.Stage;
 import java.io.IOException;
 
 public class Application extends javafx.application.Application {
+
     @Override
     public void start(Stage stage) throws IOException {
         FXMLLoader fxmlLoader = new FXMLLoader(Application.class.getResource("view.fxml"));
         Scene scene = new Scene(fxmlLoader.load());
         stage.setTitle("Electric Field Lines");
+        stage.setResizable(false);
         stage.setScene(scene);
         stage.show();
     }
