+++ src/main/java/me/petrolingus/electricfieldlines/Controller.java
@@ -6,6 +6,8 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.Pane;
 import javafx.scene.paint.Color;
+import me.petrolingus.electricfieldlines.core.DataGenerator;
+import me.petrolingus.electricfieldlines.core.Triangulation;
 import me.petrolingus.electricfieldlines.util.Point3d;
 import me.petrolingus.electricfieldlines.util.Triangle;
 import org.apache.commons.math3.geometry.euclidean.threed.Vector3D;
@@ -16,14 +18,12 @@ import org.jzy3d.colors.colormaps.ColorMapRainbow;
 import org.jzy3d.javafx.JavaFXChartFactory;
 import org.jzy3d.maths.Coord3d;
 import org.jzy3d.plot3d.builder.Builder;
-import org.jzy3d.plot3d.primitives.CompileableComposite;
 import org.jzy3d.plot3d.primitives.Shape;
 import org.jzy3d.plot3d.rendering.canvas.Quality;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Random;
 
 public class Controller {
 
@@ -33,34 +33,13 @@ public class Controller {
     @FXML
     private Pane plot3d;
 
-    private static final double RADIUS = 0.3;
-    private static final double SHIFT = 0.2;
-    private static final double ANGLE = 0.5;
-    private static final int N = 32;
-    private static final boolean SHOW_TRIANGULATION = false;
+    private static final boolean SHOW_TRIANGULATION = true;
 
-    private static final double COS135 = -Math.sqrt(2) / 2;
-
-    public static final List<Triangle> triangles = new ArrayList<>();
-    public static final List<Point3d> vertices = new ArrayList<>();
-
-    public double cx = 0.1;
-    public double cy = 0.45;
     double min = Double.POSITIVE_INFINITY;
     double max = Double.NEGATIVE_INFINITY;
-    double outerQ = 9;
-    double innerQ = -9;
 
-    int outerEdgePoints = 0;
-    int innerEdgePoints = 0;
-
-    Random random = new Random();
-
-    public static final List<Point3d> points = new ArrayList<>();
-
-    private ImageView imageView;
-    private AWTChart chart;
-    private JavaFXChartFactory factory;
+    public static List<Point3d> points;
+    public static List<Triangle> triangles;
 
     public void initialize() {
         generationOfPoints();
@@ -69,156 +48,21 @@ public class Controller {
         draw();
 
         // Jzy3d
-        factory = new JavaFXChartFactory();
-        chart  = getDemoChart(factory, "offscreen");
-        imageView = factory.bindImageView(chart);
+        JavaFXChartFactory factory = new JavaFXChartFactory();
+        AWTChart chart = getDemoChart(factory, "offscreen");
+        ImageView imageView = factory.bindImageView(chart);
         plot3d.getChildren().add(imageView);
         factory.addSceneSizeChangedListener(chart, plot3d);
     }
 
     private void generationOfPoints() {
-
-        double step = 2.0 / (N - 1);
-        int indexOuter135 = -1;
-        int indexInner135xLeft = -1;
-        int indexInner135xRight = -1;
-        int indexInner135yLeft = -1;
-        int indexInner135yRight = -1;
-
-        // Add inner points
-        for (int i = 0; i < N; i++) {
-            for (int j = 0; j < N; j++) {
-                double x = -1 + j * step;
-                double y = -1 + i * step;
-                double d = Math.sqrt(x * x + y * y);
-                double d2 = Math.sqrt(Math.pow(cx - x, 2) + Math.pow(cy - y, 2));
-                boolean isOuterContain = d < 1;
-                boolean isCloseToOuterCircle = Math.abs(1 - d) < (step / 2);
-                boolean isInnerCircleContain = d2 < RADIUS;
-                boolean isCloseToInnerCircle = Math.abs(RADIUS - d2) < (step / 2);
-                boolean condition = isOuterContain && !isCloseToOuterCircle && !isInnerCircleContain && !isCloseToInnerCircle;
-                if (condition) {
-                    points.add(new Point3d(x, y, 0));
-                }
-                indexOuter135 = (indexOuter135 == -1 && x > COS135) ? j : indexOuter135;
-                indexInner135xLeft = (indexInner135xLeft == -1 && x > cx + RADIUS * COS135) ? j : indexInner135xLeft;
-                indexInner135xRight = (indexInner135xRight == -1 && x > cx - RADIUS * COS135) ? j : indexInner135xRight;
-                indexInner135yLeft = (indexInner135yLeft == -1 && y > cy + RADIUS * COS135) ? i : indexInner135yLeft;
-                indexInner135yRight = (indexInner135yRight == -1 && y > cy - RADIUS * COS135) ? i : indexInner135yRight;
-            }
-        }
-
-        // Add outer edge points
-        for (int i = indexOuter135; i < N - indexOuter135; i++) {
-            double a = -1 + i * step;
-            double b = Math.sqrt(1 - a * a);
-            points.add(new Point3d(a, -b, 0, true, outerQ));
-            points.add(new Point3d(a, b, 0, true, outerQ));
-            points.add(new Point3d(b, a, 0, true, outerQ));
-            points.add(new Point3d(-b, a, 0, true, outerQ));
-            outerEdgePoints++;
-        }
-
-        // Add inner edge points
-        for (int i = indexInner135xLeft; i < indexInner135xRight; i++) {
-            double a = -1 + i * step;
-            double b = Math.sqrt(RADIUS * RADIUS - (a - cx) * (a - cx));
-            points.add(new Point3d(a, -b + cy, 0, true, innerQ));
-            points.add(new Point3d(a, b + cy, 0, true, innerQ));
-            innerEdgePoints++;
-        }
-        for (int i = indexInner135yLeft; i < indexInner135yRight; i++) {
-            double a = -1 + i * step;
-            double b = Math.sqrt(RADIUS * RADIUS - (a - cy) * (a - cy));
-            points.add(new Point3d(b + cx, a, 0, true, innerQ));
-            points.add(new Point3d(-b + cx, a, 0, true, innerQ));
-            innerEdgePoints++;
-        }
-    }
-
-    private void methodBowerWatson(Point3d point3d) {
-
-        vertices.add(point3d);
-
-        List<Integer> badTriangles = new ArrayList<>();
-        for (int i = 0; i < triangles.size(); i++) {
-            if (triangles.get(i).isContain(point3d.x(), point3d.y())) {
-                badTriangles.add(i);
-                triangles.get(i).setBad(true);
-            }
-        }
-
-        List<List<Integer>> polygon = new ArrayList<>();
-
-        for (int i = 0; i < badTriangles.size(); i++) {
-            List<List<Integer>> edges = triangles.get(badTriangles.get(i)).getEdges();
-            for (int j = 0; j < 3; j++) {
-                List<Integer> edge = edges.get(j);
-                boolean edgeIsNotShared = true;
-                for (int k = 0; k < badTriangles.size(); k++) {
-                    if (i == k) continue;
-                    List<List<Integer>> otherEdges = triangles.get(badTriangles.get(k)).getEdges();
-                    for (int w = 0; w < 3; w++) {
-                        List<Integer> otherEdge = otherEdges.get(w);
-                        boolean c1 = edge.get(0).equals(otherEdge.get(0)) && edge.get(1).equals(otherEdge.get(1));
-                        edgeIsNotShared &= !c1;
-                    }
-                }
-                if (edgeIsNotShared) {
-                    polygon.add(edge);
-                }
-            }
-        }
-
-        triangles.removeIf(Triangle::isBad);
-
-        for (List<Integer> edge : polygon) {
-            triangles.add(new Triangle(edge.get(0), edge.get(1), vertices.size() - 1));
-        }
+        DataGenerator generator = new DataGenerator();
+        points = generator.generate();
     }
 
     private void triangulation() {
-
-        Point3d p0 = new Point3d(-1, -1, 0);
-        Point3d p1 = new Point3d(1, -1, 0);
-        Point3d p2 = new Point3d(1, 1, 0);
-        Point3d p3 = new Point3d(-1, 1, 0);
-        Point3d p4 = new Point3d(cx, cy, 0);
-
-        vertices.add(p0);
-        vertices.add(p1);
-        vertices.add(p2);
-        vertices.add(p3);
-
-        triangles.add(new Triangle(0, 1, 2));
-        triangles.add(new Triangle(0, 2, 3));
-
-        methodBowerWatson(p4);
-
-        for (Point3d p : points) {
-            methodBowerWatson(p);
-        }
-
-        // Remove super-structure triangles
-        triangles.removeIf(t -> {
-            boolean res = false;
-            for (int i = 0; i < 4; i++) {
-                boolean c0 = t.getAid() == i;
-                boolean c1 = t.getBid() == i;
-                boolean c2 = t.getCid() == i;
-                res |= c0 || c1 || c2;
-            }
-            return res;
-        });
-
-        // Remove triangles in inner circle
-        triangles.removeIf(t -> {
-            boolean c0 = t.getAid() == 4;
-            boolean c1 = t.getBid() == 4;
-            boolean c2 = t.getCid() == 4;
-            return c0 || c1 || c2;
-        });
-
+        Triangulation triangulation = new Triangulation();
+        triangles = triangulation.create(points);
     }
 
     private void process() {
@@ -330,7 +174,7 @@ public class Controller {
                         Vector3D normal = v1.crossProduct(v0);
                         double ai = normal.getX();
                         double bi = normal.getY();
-                        double s =  normal.getNorm() / 2.0;
+                        double s = normal.getNorm() / 2.0;
                         value += s * (ai * ai + bi * bi);
                     }
