+++ src/main/java/me/petrolingus/electricfieldlines/util/Triangle.java
@@ -2,6 +2,7 @@ package me.petrolingus.electricfieldlines.util;
 
 import javafx.scene.paint.Color;
 import me.petrolingus.electricfieldlines.Controller;
+import org.apache.commons.math3.geometry.Point;
 import org.apache.commons.math3.geometry.euclidean.threed.Vector3D;
 
 import java.util.List;
@@ -31,15 +32,9 @@ public class Triangle {
         this.raid = aid - 5;
         this.rbid = bid - 5;
         this.rcid = cid - 5;
-        calcCircle();
     }
 
-    private void calcCircle() {
-
-        Point3d a = Controller.vertices.get(aid);
-        Point3d b = Controller.vertices.get(bid);
-        Point3d c = Controller.vertices.get(cid);
-
+    public void calcCircle(Point3d a, Point3d b, Point3d c) {
         double aa = a.mod();
         double bb = b.mod();
         double cc = c.mod();
@@ -96,62 +91,5 @@ public class Triangle {
 
     public int getRcid() {
         return rcid;
-    }
-
-    public double[] magicCalc(int index) {
-
-        Point3d a = Controller.points.get(raid);
-        Point3d b = Controller.points.get(rbid);
-        Point3d c = Controller.points.get(rcid);
-
-        if (index == raid) {
-            a = new Point3d(a.x(), a.y(), 1.0);
-            b = new Point3d(b.x(), b.y(), b.y());
-            c = new Point3d(c.x(), c.y(), c.y());
-        } else if (index == rbid) {
-            a = new Point3d(b.x(), b.y(), 1.0);
-            b = new Point3d(a.x(), a.y(), a.y());
-            c = new Point3d(c.x(), c.y(), c.y());
-        } else if (index == rcid) {
-            a = new Point3d(c.x(), c.y(), 1.0);
-            b = new Point3d(b.x(), b.y(), b.y());
-            c = new Point3d(a.x(), a.y(), a.y());
-        }
-
-        // USE FOR DEBUG
-        if (a.z() == 0 && b.z() == 0 && c.z() == 0) {
-            System.err.println("ALL TRIANGLES IN ONE PLANE!!!!!");
-            System.exit(-1);
-        }
-
-//        Vector3D v0 = new Vector3D(b.x() - a.x(), b.y() - a.y(), b.z() - a.z());
-//        Vector3D v1 = new Vector3D(c.x() - a.x(), c.y() - a.y(), c.z() - a.z());
-//        Vector3D vector3D = v0.crossProduct(v1);
-//        double S = vector3D.getNorm() / 2.0;
-
-//        double a01 = b.x - a.x;
-//        double a02 = c.x - a.x;
-//
-//        double a11 = b.y - a.y;
-//        double a12 = c.y - a.y;
-//
-//        double a21 = b.z - a.z;
-//        double a22 = c.z - a.z;
-//
-//        double A = a11 * a22 - a21 * a12;
-//        double B = -(a01 * a22 - a21 * a02);
-
-//        double A = vector3D.getX();
-//        double B = vector3D.getY();
-
-        Vector3D p1 = new Vector3D(a.x(), a.y(), a.z());
-        Vector3D p2 = new Vector3D(b.x(), b.y(), b.z());
-        Vector3D p3 = new Vector3D(c.x(), c.y(), c.z());
-
-        double A = (p2.getY() - p1.getY()) * (-1) - (p3.getY() - p1.getY()) * (-1);
-        double B = (p2.getX() - p1.getX()) - (p3.getX() - p1.getX());
-        double S = Math.abs((p2.getX() - p1.getX()) * (p3.getY() - p1.getY()) - (p2.getY() - p1.getY()) * (p3.getX() - p1.getX())) / 2;
-
-        return new double[]{A, B, S};
     }
 }
